Manage the thrift interface requiring in thrift.rb
When thrift generates the code, it inserts require statements at the top of the files
that do not work in qup. They are relative to the file, not full paths to the files
that need to be required. Therefore the loading of the thrift generated code is being delegated to the parent file 'thrift.rb'