Ensure a topic is aware of changes in subscribers
No tests existed that ensured that when two different handles to the same
topic existed, if a subscriber was added to one instance, it would show up
in the other instance. This is now fixed.

This exposed an error in the maildir adapter which would not reload the 
list of subscribers if they changed.

This exposed an error in the redid adapter which would not cleanup subscribers if the topic was deleted.